Extract ReflectiveElementHandler
