+++ kotlinpoet-km-specs/src/main/kotlin/com/squareup/kotlinpoet/km/specs/JvmDescriptorUtils.kt
@@ -2,9 +2,6 @@ package com.squareup.kotlinpoet.km.specs
 
 import kotlinx.metadata.jvm.JvmFieldSignature
 import kotlinx.metadata.jvm.JvmMethodSignature
-import java.lang.reflect.Constructor
-import java.lang.reflect.Field
-import java.lang.reflect.Method
 import javax.lang.model.element.Element
 import javax.lang.model.element.ExecutableElement
 import javax.lang.model.element.NestingKind
@@ -174,67 +171,3 @@ internal object JvmDescriptorTypeVisitor : AbstractTypeVisitor6<String, Types>()
 
   override fun visitUnknown(t: TypeMirror, types: Types): String = error("Unsupported type $t")
 }
-
-internal val Class<*>.descriptor: String get() {
-    return when {
-      isPrimitive -> when (kotlin) {
-          Byte::class -> "B"
-          Char::class -> "C"
-          Double::class -> "D"
-          Float::class -> "F"
-          Int::class -> "I"
-          Long::class -> "J"
-          Short::class -> "S"
-          Boolean::class -> "Z"
-          Void::class -> "V"
-          else -> throw RuntimeException("Unrecognized primitive $this")
-      }
-      isArray -> name.replace('.', '/')
-      else -> "L$name;".replace('.', '/')
-    }
-}
-
-internal val Method.descriptor: String get() {
-    return buildString {
-        append('(')
-        parameterTypes.joinTo(this, transform = Class<*>::descriptor)
-        append(')')
-        append(returnType.descriptor)
-    }
-}
-
-/**
- * Returns the JVM signature in the form "$Name$MethodDescriptor", for example: `equals(Ljava/lang/Object;)Z`.
- *
- * Useful for comparing with [JvmMethodSignature].
- *
- * For reference, see the [JVM specification, section 4.3](http://docs.oracle.com/javase/specs/jvms/se7/html/jvms-4.html#jvms-4.3).
- */
-internal val Method.jvmMethodSignature: String get() = "$name$descriptor"
-
-internal val Constructor<*>.descriptor: String get() {
-    return buildString {
-        append('(')
-        parameterTypes.joinTo(this, transform = Class<*>::descriptor)
-        append(')')
-        append('V')
-    }
-}
-
-/**
- * Returns the JVM signature in the form "<init>$MethodDescriptor", for example: `"<init>(Ljava/lang/Object;)V")`.
- *
- * Useful for comparing with [JvmMethodSignature].
- *
- * For reference, see the [JVM specification, section 4.3](http://docs.oracle.com/javase/specs/jvms/se7/html/jvms-4.html#jvms-4.3).
- */
-internal val Constructor<*>.jvmMethodSignature: String get() = "<init>$descriptor"
-
-/**
- * Returns the JVM signature in the form "$Name:$FieldDescriptor", for example: `"value:Ljava/lang/String;"`.
- *
- * Useful for comparing with [JvmFieldSignature].
- *
- * For reference, see the [JVM specification, section 4.3](http://docs.oracle.com/javase/specs/jvms/se7/html/jvms-4.html#jvms-4.3).
- */
-internal val Field.jvmFieldSignature: String get() = "$name:${type.descriptor}"
